requirements.txt
numpy
pandas
scipy
scikit-learn
pathlib
dataframe-image
distinctipy
matplotlib
umap-learn
seaborn
datetime
